requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
celery==5.6.2
celery-types==0.24.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==6.0
django-stubs==5.2.8
django-stubs-ext==5.2.8
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
iniconfig==2.3.0
kombu==5.6.2
networkx==3.6.1
numpy==2.4.1
packaging==25.0
pluggy==1.6.0
prompt_toolkit==3.0.52
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
pytest-django==4.11.1
python-dateutil==2.9.0.post0
redis==7.1.0
six==1.17.0
sqlparse==0.5.5
types-networkx==3.6.1.20251220
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.2.14
